Thanks to Marcel Wiesweg <marcel.wiesweg@gmx.de> for the following patch:
when I started to work with KimDaBa, I noticed first that it is an excellent
piece of work, and second a few major performance hits:

1. Thumbnail creation. A 64x64 thumbnail from a 3.2 Megapixel image took ~1.2
seconds on my machine. Imagine me sitting in front of 1000+ images to
classify...

2. The "one-at-a-time" ImageConfig dialog took veeery long to appear (3 s)

3. Switching to the next image in the same dialogue also took a good time.

I traced all the problems in last consequence down to the thumbnail creation.
The patch attached does the following:

- make use of libjpeg and its scale-while-decoding capability wherever
possible (now thumbnail takes ~0.2s => factor 6)

- use the excellent ImageManager/ImageLoader infrastructure in the ImageConfig
dialogue - if necessary, the preview appears after a short delay

- use a preloading mechanism in this dialogue: The next image is preloaded,
the last is kept in cache. This has the effect that in 'normal' operation the
preview always appears instantly.

svn path=/trunk/kdeextragear-2/kimdaba/; revision=320309