My dad just died 77 years old, yesterday - I will always love you, and never forget you.
During his last days I was sitting next to his bed comforting him, and
while doing so I was (as always) working on KPhotoAlbum, so here comes a
massive checking.

- renamed Movie to Video throughout the code.
- replaced the listbox in the annotations dialog with a listview, which now
makes it possible to implement two long awaited features.
- first feature is already implemented, namely removing non matching items
while typing in the line edit. The second feature which will be implemented
next is to show categories in the list view.
- now it is also possible to press arrow up and down in the line edit in
the annotation dialogs, and that way browse the items matched so far during
typing.
- fixed some signal/slot connection problems that was induced when adding namespaces.

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=562760