Restructurize SQLDB category class hierarchy.
Point of all this is to make it possible create special categories
like Folder and Media Type, which don't use same methods for
classifying or searching. This is because special categories don't use
category and tag table like normal categories do.

Now there is a classify method in SQLCategory, which should be
overwritten by concrete categories.

Entry named Folder will not be created to category table anymore.
(There shouldn't be one, because it doesn't have any tags.)

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=600898