How very rare that you (1) get a new feature to work (2) improve performance significantly and (3) get much cleaner code base ALL three at the same time :-)
I beleive that pixel by pixel zoom now works in the reworked version. As an added bonus, I no longer loads the first image in full size, but
only in viewer size. This has given a significant boost in starting up the viewer.

If the user ever zooms, then the real sized image is loaded in the background and put in place when it is available - until then the user just sees
the viewer sized version he was original looking at.

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=602827