Don't try to parse & and | directly from string query, but use different API where we pass an QStringList. Each item in this list is either "condition" or "operator". This way, we can now browse directories with names containing funny special characters like & or |.
This fix is not complete because it still isn't possible to search for those
special names via Search dialog.

I'm afraid this will have to be revisited at some point in future because it
still doesn't allow for more complicated queries like "(a & b ) | (c & d)"
without relying on operator precedence.

BUG: 147871

svn path=/branches/extragear/kde3/graphics/kphotoalbum/; revision=726066