Preload image thumbnails from images in the thumbnail viewer even if they aren't visible yet.
Preloading starts as soon as the ThumbnailWidget knows what images it contains
and continues until another bunch of images is to be shown. This behavior is
open for discussion, perhaps it should stop when ThumbnailWidget goes out of
(visible) place, perhaps it should only load images that aren't too far from
current viewport etc.

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=832284