Pass lists as a single QVariant in the bindings
The previous usage was an error, because the list was not wrapped as
single QVariant but instead appended to the bindinds list directly,
and so the other end (in QSqlConnection) was not able to parse the
list parameters. Now the other side can replace the place holder
(i.e. question mark) of the list with the actual contents of the list.

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=856314