Refactor DB::Result to be implicitly shared
There is no need for shared pointer wrapper ([Const]ResultPtr) any
more and that makes things easier when passing around DB::Results.

This change also makes it possible to use Q_FOREACH with DB::Result,
which makes iterating syntax easier to read.

The main change is in DB/Result.h and other changes are made to match
that API change.

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=870572