Viewer::VideoDisplay: Do not leak Phonon objects
Each invocation of viewer with a video will create a new VideoDisplay
and it used to create Phonon::MediaObject and Phonon::AudioOutput
without a parent, but it did not delete those in the destructor
either, so they were leaked. That resulted in many trashed threads, if
many videos were shown.

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=936373