From Andreas:
Since the two parts "detecting new file versions" and "copy and rename
on open" in the settings dialog have nothing in common conceptually
(although they can be
used in tandem quite well) I propose separating them in the options
dialogue a little more.

Thanks to Andreas Neustifter for the patch.

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=1137918