Preload thumbnails in the current scope of the thumbnail viewer
The switch to Qt's MVC took away the preloading of thumbnails that are part of
the current view, but remain invisible simply because they haven't been scrolled
to yet. This patch simply preloads all images in the current scope when the
model is set up with the list of images.

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=1184856