Rename DB::ResultId to DB::Id and DB::Result to DB::IdList
These are better names, since Result/IdList is not always a search
result but just a list of Ids. It is already used in places where it
is not a result of a search, e.g. list of ids in the same stack.

This eliminates couple old todos (as can be seen by grepping todo or
TODO from the diff).

svn path=/trunk/extragear/graphics/kphotoalbum/; revision=1210459