Do not preload files that has already been loaded This is the setup: The thumbnail view is resized, and as a consequence all thumbnails are scheduled to be loaded. At the same time, the thumbnailview will request all images visible to be created (this with a higher priority).
This commit makes the code which preloads the thumbnails check if they
are still needed (and thus not recreate them if they were already
created by the thumbnail view).

This will speed up things if there esp. is many videos visible in the
view.

Notice the above also applies to images that was scrolled too before
they were reached by the preloader.