Improved handling of events in date edits in annotation dialog
In the annotation dialog the linedit for date supports that you e.g.
just type 1979. This text is expanded to "jan 1st 1979" when you press
enter in the line edit, and it should also do so when the focus leaves
the lineedit. There was code in place for this, which relied on a focus
out event send to the line edit. We, however, never saw that event. So
now it is rewritten to used the editFinished signal from the line edit
instead, which is cleaner anyway.