Do not load video presnapshots on the main event loop
As an optimization, when a snapshot is loaded from a video file, it is
saved in a image file in ~/.videoThumbnails. When the grid was resized
that snapshot was used instead of loading a new snapshot from the video
file. Until now, that snapshot was loaded and rescaled on the GUI
thread, which made the UI freeze if there where a large number of video
files.