Add option to hide histogram.
Since I do not use the histogram it takes over precious screen real estate.
This patch adds the option to remove the histogram.

Properly hide and show histogram instead of not creating it.

The previous attempt to hide the histogram was to not create it entirely,
not its only show()n and hide()n as needed.

When the ShowHistogram setting is changed, change the visibility immediately.