Show count of selected items in AnnotationDialog.
The ListSelect widgets now change the title of their parent QDockWidget,
to include the number of selected items.

Depending on the mode of the ListSelect the Title is changed as follows:
InputSingleImageConfigMode -> "Title (%1)"
InputMultiImageConfigMode  -> "Title (%1-%2)"
SearchMode                 -> "Title"

With %1 being the number of selected items, and %2 being %1 plus the
number of unchanged("partially selected") items.