AnnotationDialog::hasChanged(): remove unneeded check.
When the Annotation dialog is closed by closing the window (not pressing
cancel), firstDescription is empty, thus marking the image as dirty.

In InputSingleImageConfigMode mode, writeInfo() already makes sure that
the image description is saved (and therefore not mised by
hasChanges()). Therefore the firstDescription stuff can be removed for
that mode.

BUG: 303569