Rewrite negation in CategoryMatchers.
Until now, negation of matchers was coded into the matcher itself,
by the way of a "Matcher sign".
This patch introduces a general "NegationCategoryMatcher", which
simply negates the outcome of a child matcher. This way, matching code
in ImageSearchInfo can be written in a more comprehensible way.

Also, the new Matcher is used to fix the behaviour of matches like "!
None".

bug: 304801