fixed ugly crash - well worked around it
Here is the full storry.
NewImageFinder::loadExtraFile is searching for files which are edited versions of other files.
It does so by matching the file name of the edited file with a patter.
E.g. foo-edited.png might come from foo.png
Now if both files has just been copied into the database, and it finds the foo-edited.png BEFORE
it finds foo.png then it will try and look up the image info for foo.png and fail.