Don't use singleshot timer for scanning for images
In commit 45a2fe888719f2b67570f6a51da1d9bb12a172f7 made by jkt back in 2008 he wrote:
load images on background during startup

However, that is not what will happen. What will happen is that next time the event loop runs
(when the window is shown) the UI will block while it scans for images.

I'm sure this was not his intended behavior. Esp because the spashscreen actually tries
to notify the user that we are scanning for images, but with this delayed,
this message is never shown.