Remove special treatment of deprecated standard categories.
"Peoples", "Locations" and localized category names in the DB (as far as
we can catch them) are now automatically converted to their modern
counterpart. I.e. we don't need any special treatment for them in the
rest of the code.

"Keywords" is still treated a standard category, since "Events" is not a
1-to-1 equivalent.

The database version is also pushed by this change, because the
conversion should be a onetime operation only.