Add class Map::SearchMarkerTiler.
This class is introduced so that we can influence the return value of
KGeoMap::ItemMarkerTiler::getGlobalGroupState() in a way that enables
the clearSelection button in the MapView in search mode.

Actually, we always report a region selection now, but that should not
matter since the selection is only enabled in search mode.