Reserve one core for main thread.
The previous code uses the number of CPUs (at least one),
but usually all cores to use MPlayer to generate video thumbs.

This brings a dual core machine (spaning two MPlayers) to its knees, so
in the patch I have "reserved" one core (using only (n-1) for MPlayer)
for running the main KPA app. This will not be as noticable on quad- and
octa-cores but still KPA will run more smoothly on those machines.