CompletableLineEdit: fix up/down key for single match.
This change fixes the behaviour of the up/down keys when there is only a
single match. In that case, the old behaviour was to not select anything
in case the arrow keys were used.
With the new behaviour, the down key starts selecting at the first
line, instead of the second line.