Fix crash when clicking items in browser.
Thanks to Matthias Füssel for sending this patch!

If you connect the "itemClicked" slot with the "activated" signal
instead of the "clicked" signal (see attached diff), clicking cateories
and tags works again. You also get the "selected item will be triggered
when the return key is pressed" behaviour for free (i.e. without the
code in the event filter) if you use "activated".