Use semi-manual versioning.
Since current versions of releaseme (and also the workflow in neon)
don't allow for insertion of a version.h header file at tarball
creation, manual versioning is now supported again.
For git versions, we still use the info from "git describe", but in a
tarball we use the version from the "project()" command in cmake.