Disable face recognition and detection by default
The feature is still available, but users have to enable the option
"ENABLE_PLAYGROUND" to build it.

This was done primarily because of the limited usefulness of the feature
in its current form. While some issues with the workflow could be fixed,
it does not make much sense to do so unless performance of the
underlying face detection library improves.