Reuse numerical tag ids.
This patch makes KPA distinguish between tag order in the xml file and
tag id. In most cases, tags will now get the same id when saving as they
had when the db was loaded from disk.
This allows us to reorder tags (as required by the SortLastUsed option
of the ListSelect) and still have diffable xml files.

CCBUG: 370373