Fix remove tokens dialog.
On KDE 5 in current KPhotoAlbum master, the "Remove Tokens" dialog shows
nothing but the "Ok"/"Cancel" button bar. With this fix it works again.
The fix also adds an ellipsis to the "Remove Tokens" menu entry,
suggesting that it will show a dialog before doing anything and so
making it appear far less scary...

Thanks to Matthias Füssel for the patch!