Add GeoPosition page in browser.
This patch creates a "map browser" similar to the category browser and
folder browser. This browser shows a map with all images that match the
current filter criteria. When you select a region, it adds another
filter that matches all images that have a geolocation and are inside
the selected area.

Thanks to Matthias Füssel for the patch!