Prevent double delete.
In ResizableFrame, a AreaTagSelectDialog is allocated and then used as
context menu. The dialog was incorrectly guarded by a QScopedPointer,
even though it has the ResizableFrame as QObject parent and is therefore
deleted by when the frame is deleted.
When the contect menu was open and the annotation dialog was switched to
the next image (e.g. using PgDown), the ResizableFrame would delete the
dialog before the scoped pointer, leading to a double delete.