Speed up removal of images from database.
Greatly speed up removal of images from database by using a transaction
for the EXIF data and effectively the same thing for the thumbnail
cache.

Both of these operations are currently performed one file at a time
rather than as a transaction.  The case of re-creating the EXIF database
was resolved with commit 62b2060 in January of last year; this resolves
the same issue for deleting files.

There should be some kind of error handling for removal from the EXIF
database, but I'm not sure exactly what's best.  Failure to delete a
record because the record isn't already present should not be an error.