Improve robustness of video length detection and thumbnail extraction.
1. length is determined from the container as some vstreams do not have
the attribute duration (as You suggested).

2. verbosity is set to 0 (panic).

3. We just want a number and nothing breaks if there is junk in the
first line of output (this is caught later on where the cast to double
is checked). Therefore, the test is: "do we have at least 1 line"

4. For thumbnailextraction I need the analyzeduration parameter, even if
it slows things down a bit. With some videos we don't get the thumbs
otherwise.
It would be better to *try* without this parameter first, then *retry
with* on failure and keep the parameter for this specific file. But I
did not find an easy way to retrace my steps and try the same file twice
as these processes run asycronously.