Use uppercase tokens to add token to whole stacks.
This patch lets
you use <shift>-tag to apply the tag to (or remove it from) all
members of stacks containing the selected images, not just the images
that are currently selected.

My most common use case for this is if I'm shooting RAW+JPEG, and I
quickly reviewing the images to decide which ones to select.  So my
first action is to autostack the new images, and then review them and
tag the ones I want.  Normally this will only tag the top of the stack
(the one I'm seeing if stacks are collapsed), making it somewhat
painful for me to easily get to all of the RAWs or all of the JPEGs of
the ones I've tagged.  Using uppercase tags lets me select by tag to
get all members of each stack.