Fix compilation on FreeBSD.
OptimizedFileList uses statfs, which isn't standardized between Linux
and BSD. There's also a statvfs function which is standardized, but that
"is not meaningful" in the FreeBSD implementation.
OTOH, our usage of statfs does not use other, non-compatible parts of
statfs, so we get away with adjusting the includes for the respective
platforms.