Fix zooming for raw images.
The ImageDisplay falsely assumed that after requesting a full size image
the resulting image has the size that is stored in the ImageInfo.
Usually this is true, but when the embedded preview of a raw image is
used, the preview often has a different size as the full raw image.

Therefore, even though we just loaded a "full size" image, it has a
smaller size than the raw image.