Recover from duplicate entries in index.xml
KPhotoAlbum assumes filenames to be a unique identifier of images (there
can not be two different files using the same file name).
This patch automatically recovers if somehow an entry in index.xml was
duplicated, merging both entries into one. If the MD5 sum of both
entries does not match, an error is raised and KPhotoAlbum exits.