Classify categories by item count & date range.
Instead of only counting items in a category, also take not of the range
of dates in the category when classify()ing.
This information is then used to add a date range column to the
CategoryPage.
Note: currently, the date range is only calculated for "plain"
categories. Groups are ignored for now.