Use sqrt instead of logarithm for histogram scaling.
As mentioned by Robert, this should be more intuitive than logarithmic
scaling.
Also bundled in this commit (since it doesn't make sense to split this
out): inverted wording in the settings dialog to "Use linear scaling for
histogram". With this wording we can tweak the exact scaling method
without having to come up with a new settings name every time.

CCBUG: 356810