Flush name cache when changing database compression.
For performance reasons, FileReader::unescape and FileWriter::escape
cache the results of (un)escaping.
With the recent (introduced in v5.4) push to use the compressed file
format by default, kphotoalbum would load an uncompressed file and later
store it as a compressed one. In this process, the cache would be
populated for an uncompressed file and later used to store a compressed
file, producing incorrect results.
On the next attempted database load, these incorrect results may be
discarded if they are invalid XML attribute names, thus leading to data
loss.

BUG: 403668