Add DB::UIDelegate.
This class allows us to break the dependency on MainWindow that we
currently have in DB and XMLDB in the form of the messageParent()
methods in XMLDB/FileReader and XMLDB/FileWriter.

So far, thie status quo is untouched - we still need to make a non-dummy
UIDelegate and then move the KMessageBox usages over to UIDelegate.

I'm not yet sure if we also need to break the dependency in
NewImageFinder or if we just see it conceptually as part of the GUI.