Implement UIDelegate interface in MainWindow::Window.
The current approach of using a messageParent and then calling UI
methods directly from the DB code also uses MainWindow as a message
parent for the dialogs, so this is the obvious choice for a ui delegate.

The MainWindow has the same or longer lifetime as the database, making
it a good choice for the delegate.
We could also introduce a different class that only provides the glue
code between a QWidget as message parent and the UIDelegate, and this
class would only take a few bytes and be more flexible. OTOH this
flexibility is not needed and the few methods for the new interface
won't hurt the main window.