Show new areas in annotation preview.
Show the new/altered areas as highlighted in the fullscreen preview of
the annotation dialog.

Previously, the fullscreen preview would only show the tagged areas that
were already set when opening the annotation dialog.
The new behavior is to additionally show the newly created areas in a
different color so that is easy to distinguish old from new areas.