WIP: Filter thumbnialview by rating/token.
This does not work in its current form, because I'm running out of
viable keyboard modifiers. I had planned on:
 - Ctrl + (token or rating): set filter
 - Escape: clear filter if set.

This does not work because Ctrl-1-4 are already used for the annotation
dialog and stack maipulation, and Ctrl+[a-z] has plenty of collisions
with existing keyboard shortcuts, obviously.
The shift modifier is also used in the KeyboardEventhandler (affect
stack vs only visible item).
The meta modifier has the same problems as Control.