Remove thumbnail filter by rating.
Doing this directly in the keyboard-handler IMO isn't worth it - we can
add a dedicated toolbar for this later. The toolbar can have
user-configurable accelerators, which is a win-win.

As for the tokens: this can only realistically be handled here, because
potentially setting 26 shortcuts won't be a nice user experience.