Fix handling of raw image rotation.
Raw images usually contain an embedded jpeg preview image that is
sufficient for many use cases.
The embedded preview is returned by libkdcraw in its non-rotated form
(probably not to confuse renderers, since the preview image has exiv
orientation data), while the raw image is returned in its rotated form.
Both images have until now been passed through the same pipeline as all
other images, rotating (and scaling) the image before display.

This leads to an incorrectly rotated image as soon as the raw image is
decoded (because it is rotated a second time). This patch bypasses the
rotation when the source is a decoded raw image.

BUG: 412192