Map: use pins by default.
Performance-wise, this makes a huge impact. Fast should be the default,
pretty is secondary ;-)
On my not-so-fast computer, showing pins for ~5.5k images with
geoinformation takes ~20ms vs ~6.5s for thumbnails.

A nice further tweak would be to check whether a thumbnail actually
needs to be rendered or not. An even nicer thing would be some
level-of-detail or automatic clustering of thumbnails in proximity of
each other.