WIP: Draw GeoClusters in screen coordinates.
This should allow for a smooth transition when zooming out, i.e.
instead of switching from geo coordinates when we're close up to screen
coordinates when the area wouldn't be visible otherwise the new code
calculates the screen size of an area and then draws the area in screen
oordinates.
The problem so far: screen coordinates seem to be computed incorrectly.
FIXME: clang-format