Disable build of remote client by default.
The install base of the android app is arguably zero at this point (the
app hasn't been available in any app store for some time now), and
people seem to be confused because they can't find the app.

So instead of getting people's hopes up I decided to disable building
the remote control capabilities for now.

I see two possible future scenarios:
a) Someone picks up work on the companion app and we reinstate it.
b) We remove the app completely, probably some time in 2020.

If we pick scenario a), we need to integrate this with the KDE CI/CD
infrastructure and have the publishing handled by KDE. I don't want
publishing in the Play Store to depend on a single person like in the
past.