Exif search: remove search by aperture.
We allowed both searching by the APEX aperture value and by the F
number, but the APEX value is very rarely used in photography and most
cameras don't even record it. In addition to that, the values for the
range selection are set to be identical than their F number
counterparts, which is incorrect.

Thus, I remove the option that's plain confusing and hardly useful in
its current form.