Fix crash when associating a tag with an area.
If a tag is being associated with an area, the context menu has two
levels: the direct "associate with <last added tag>" and the sub-menu
"associate with..." followed by a list of tags.

Each version uses a different function signature for
ResizableFrame::associateTags. When refactoring connections to the "new"
signal-slot syntax in commit c38f5136 (thus released in version 5.6), I
applied the wrong signature to the connect statement.

This resulted in a crash/failed assertion when selecting a tag in the
described way.