Consistently show DeleteDialog when closing viewer.
If deleting images in the viewer using <Ctrl+Del>, the images are marked
for deletion and a MainWindow::DeleteDialog is shown when the viewer is
closed.
Until now, the dialog would only be triggered when closing the viewer
using <Esc> or via the context menu, but not when closing the window
directly.