Fix crash when images are deleted while the viewer window is open.
When images that are in the ViewerWidgets' playback list are deleted
outside the viewer, it was possible for the viewer to try to access the
deleted image in the database.
Fortunetely, the ImageDB already provides a signal that can be used to
remove the images from the playback list once deleted.