Fix crash when changing the toolbar.
If anything in the application toolbar is changed (new item, renaming
items, etc.), the search bar is recreated to avoid inconsistencies.
As a side effect, this meant that the filterWidget was deleted, leading
to ThumbnailFacade::filterWidget returning an invalid pointer.

Therefore, ThumbnailFacade needs to create a new FilterWidget every time
one is requested. As a side effect, ThumbnailFacade::actions needed to
be removed (there is no permanent FilterWidget object anymore).