Remove password pseudo-protection from privacy lock.
The privacy lock is a feature in KPhotoAlbum that allows the user to
"lock away" some parts of their database so that one doesn't show
sensitive images to others by accident.

Until now, it was possible to optionally set a password for
disabling the lock.
The privacy lock always came with a big warning informing the user that
this is no real protection since the password is easily looked up from
the configuration file.
To make the intention of the privacy lock clearer, from now on it is
only possible to lock/unlock the images, without any password prompt.