ThumbnailCache: store thumbnail size in cache index
The way I intend this is to always use the stored thumbnail size from
the thumbnail cache and set the settings value from this.
Invalidating the thumbnail cache is now done slightly different: the
ThumbnailCache now does this automatically when needed.

For the rest of KPA, SettingsData::thumbnailSize() is still the main
source of information, and enforcing size constraints is also done in
the settingsdata.

This prevents inconsistent thumbnail data from being generated when the
config file was edited by any outside program.