Break dependency FileName -> ImageDB (Part 3/3)
Remove FileName::info().
This also brought out lots of usages that qtcreator didn't find earlier.

For ImageDB::count() it made sense to also change the ImageDB::search()
method to return a ImageInfoList instead. Creating an ImageInfoList
takes actually less work for search(), and the list is easily converted
to a FileNameList when needed.