SettingsData: remove dependencies on QApplication and KMessageBox.
Using UIDelegate works to remove the dependency of SettingsData on
having a GUI application for now.
Probably a better solution would be to rethink the whole "trust
timestamps" setting. In 2020, having the "Ask" option seems like a
strange notion to me. Removing the option, though, is probably more
hassle than removing the whole trust timestamps setting.