Introduce libraries kpabase and kpathumbnails
This is a first step towards isolating different parts of kphotoalbum
into functionally independent libraries, making it easier to do
refactorings in the future.

KPA::Base
-> common base parts with minimal dependencies.
-> basically KPA's "tier 0"

KPA::Thumbnails:
-> access to the thumbnail cache
-> maybe access to thumbnail building in the future (if separable from
GUI code)

Going with this theme, Exif and DB may be the next libraries to come,
but for now I'm content with a minimal set of libraries required to get
the thumbnailtool on track.