Consider MemberMap data when saving categories.
If a tag group is only known in the MemberMap (because it was not used
as a tag on an image or video), then special care is needed when saving
to prevent inconsistent data from being written. See the attached bug
number for an analysis of a possible bug scenario.

BUG: 423334