Start fixing test failures introduced by 185b1344b1fba4f61d91f91304c49b787d891494
When I fixed the MD5 sums in the demo database in April, I didn't pay
attention to the consequences for the integration tests that this would
cause.
Mainly, the MD5 sums also need to be fixed in all test cases. This is
done in this commit.
Additionally, the fixup in April introduces some errors into the demo
database: angle="x" was removed erraneously even where x != 0 (this is
also fixed by this commit), and timestamps were changed incorrectly.
This is also fixed.

At this point, though, only 1/8 tests succeeds - more investigation is
needed.