Fix inconsistencies when changing color schemes.
Instead of changing the palette of a widget, it is often possible to
change the foreground and/or background role to use different colors of
the standard palette. This way, changes in the palette can propagete to
the widget without additional work.

When this is not an option, the widget needs to react to palette changes
explicitly and (re-)adjust the palette accordingly.

UI changes: The exif grid now uses regular highlight colors instead of
hardcoded bright red.

A very big thank you to David Redondo for helping me get a better
understanding of the palette system and for pointing out how to avoid
breaking palette propagation!