Add SPDX headers for CMakeLists.txt
So far, there was no explicit license set for the CMakeLists.txt files.
Yet, until I recently started the conversion to SPDX headers we had the
COPYING-CMAKE-SCRIPTS file, which implies a BSD-3-Clause license for all
cmake script files.

As far as actualy copyrights go, I assumed any git commit author
contributed something to the files (which is a big oversimplification,
but for the build files I'm willing to lean out of the window a little
to have a clean starting point).