Add license for pre-commit hook
This file is the default file that comes with every git repository. In
the git source code, this file corresponds to
templates/hooks--pre-commit.sample

This file does not have an explicit statement, and in git's own
repository I could not find anything definite (beyond the license for
the complete product). From this, and the Debian copyright file
(/usr/share/doc/git/copyright) I conclude that GPL-2 is indeed the
correct license:

Files: *
Copyright: © 2005-2020, Linus Torvalds and others.
License: GPL-2