Refactor ExternalPopup
The current code loads a list of KServices, saves only their
user-visible names and icons and later reconstructes the KService via
KMimeTypeTrader. Instead keep the KService itself around so we can use
it later.

The code also connects to the triggered signal of the menu and then uses
the user-visible action name to determine what codepath to trigger.
Instead connect to the triggered signal of the individual actions with
slots for the specific actions to better separate the codepaths.