Add cache for scaled thumbnails in MapView.
On my computer, caching the scaled pixmap in the GeoBin instead of using
the regular thumbnail cache dropped the render time per image from
around 1.25ms to virtually nothing (~200 thumbnails now use 0ms instead
of ~250ms).

Given that the marker size is 40x40 pixels and the cache is only
populated if a GeoBin is actually rendered, I estimate the worst-case
memory consumption to be acceptable.