BreadcrumbViewer: don't crash if updateText is called w/o breadcrumbs
BreadcrumbViewer::updateText() is only called when the window is not
wide enough to hold the full breadcrumb texts.
If, however, the main window is very narrow on startup, the text is
called at a time when there is no breadcrumb in the breadcrumb list,
leading to updateText accessing an invalid index of the list.