Have windows be maximized by default.
Instead of the very outdated, hardcoded size of 800x600 for the main
window and annotation dialog, don't set a default size.

This means that the annotation dialog size is now computed according to
its contents, and the main window is maximized if no size is set in the
config file.