BrowserWidget: select first item when typing in the search bar.
The behavior when typing into the search bar and then pressing enter is
that:
(a) if a selection was made (e.g. by pressing arrow keys) it is chosen
(b) if no selection was made, the first item is chosen

Case (b) did not offer any visual feedback as to which item was chosen.
By selecting the first item while typing (provided nothing was already
selected), the behavior becomes clearer to the user.