BrowserWidget: select first item as soon as SearchBar is activated
As a followup to 7a10e4bb, make sure that the first item of the active
BrowserWidget view is already selected when the SearchBar gets focus: as
soon as the SearchBar has focus, the behaviour when pressing enter has a
slightly different semantic (i.e. selecting the first item in the view
if nothing was explicitly selected as opposed to doing nothing when the
enter is pressed without any item being selected).