Browser: use activated() signal instead of clicked()
Before commit 19d42946, we would only activate items in the overview
page when clicking them, not by pressing enter on the keyboard.
The mentioned commit was intended to implement the keyboard activation
in addition to the single-click behaviour enforced by clicked().

Unfortunately, this leads to double activations because both clicked()
and activated() are emitted when presing an item.

Therefore, we now use only acivated() which may be an inconvenience for
users who use "double-click to select files/folders".