WIP: implement a ThumbnailCache::vacuum() method.
So far, I think the method should work, but it needs a test and is not
yet accessible in kpa-thumbnailtool.

The approach to vacuuming the cache is quite simple:
1. move the cache files out of the way
2. flush the cache
3. read the raw thumbaildata from the cache files and insert it using
regular functionality

The cache is not locked during step 3, so if there's anything else modifying
the thumbnail cache this is possible. As far as I can tell, this
shouldn't be a problem.