Use compile time checks instead of dead code in ViewerWidget
initiateVideoBackend can on paper return a nullptr, which would invite
all sorts of problems. In practice, however, the build system prevents
this from happening.
Thus, adding code to prevent the nullptr (as in commit
9a942aa7a9cab62b648a8c0c895811efb3fc4c31) is the wrong thing to do.
Instead, use a static_assert to document that the function never returns
a nullptr (and to safeguard against future errors in the build system).