Fix crash when returning to thumbanil view and video data come in
This occoured when the app went back to the thumbnail screen, which
resulted in pages from the listview being deleted. If those pages had a
video being loaded a cancel request was sent to the server.

However, the client would go away right away, and the server would send
the video block it was currently loading over, accessing a deleted
client.