Change Changelog format
Keepachangelog.com style changelog feels much less restricting to write,
and at the same time improves structure by recommending a set of
categories to group change entries by.

This is also a good opportunity to explicitly state the license for the
changelog to be CC-BY-SA-4.0. I have mailed all authors with noteworthy
changes to the file for permisssion to do so.