Make "display labels/categories" accessible via "view" menu.
Until now, the two settings "display labels in thumbnail view" and
"display categories in thumbnail view" were only changeable via the
settings dialog.

This commit adds them to the "view" menu to be easier to reach. As a
side affect, this change also allows both actions to be assigned a
keyboard shortcut.

BUG: 145346